amphibian-alchemists/requirements.txt
aiohttp==3.5.4
alembic==1.3.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
discord.py==1.2.5
docutils==0.16
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Kivy==1.11.1
Kivy-Garden==0.1.4
Mako==1.1.1
MarkupSafe==1.1.1
multidict==4.7.4
oauthlib==3.1.0
py-enigma==0.1
pycryptodome==3.9.4
Pygments==2.5.2
python-dateutil==2.8.1
python-dotenv==0.10.5
python-editor==1.0.4
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
SQLAlchemy==1.3.12
typed-ast==1.4.1
urllib3==1.25.7
websockets==6.0
Werkzeug==0.16.0
yarl==1.4.2
